navbar + i18n + darklight mode part 2

+++ Formula10/src/main/java/hu/project/formula10/dto/StandingDTO.java
@@ -14,5 +14,5 @@ public class StandingDTO {
     private Long groupId;
     private Long seasonId;
     private Long userId;
-    private Long totalPoints;
+    private int totalPoints;
 }
